Split single api.py into resources/*
